# openapi/adminusers_spec.yaml
openapi: 3.0.1
info:
  description: Adminusers API
  title: Adminusers API
  version: v1.0.0
paths:
...
  /v1/api/invites:
    get:
      operationId: getInvites
      parameters:
      - example: ahq8745yq387
        in: query
        name: serviceId
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invite'
          description: OK
      summary: List invites for a service
      tags:
      - Invites
  /v1/api/invites/service:
    post:
      operationId: createServiceInvite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteServiceRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
          description: Created
        "403":
          description: The email is not an allowed public sector email address
        "422":
          description: Missing required fields or invalid values
      summary: Creates an invitation to allow self provisioning new service with Pay
      tags:
      - Invites
  /v1/api/invites/user:
    post:
      operationId: createUserInvite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
          description: Created
        "404":
          description: Service or role not found
        "422":
          description: Missing required fields or invalid values
      summary: Creates an invitation to allow a new team member to join an existing
        service.
      tags:
      - Invites
  /v1/api/invites/{code}:
    get:
      operationId: getInvite
      parameters:
      - example: d02jddeib0lqpsir28fbskg9v0rv
        in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
          description: OK
        "404":
          description: Not found
      summary: Find invite for invite code
      tags:
      - Invites
    patch:
      operationId: updateInvite
      parameters:
      - example: d02jddeib0lqpsir28fbskg9v0rv
        in: path
        name: code
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
              example:
              - op: replace
                path: telephone_number
                value: +441134960000
              - op: replace
                path: password
                value: a-password
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
          description: OK
        "400":
          description: Bad request
        "404":
          description: Invite not found
        "422":
          description: Missing required fields or invalid values
      summary: Update an invite
      tags:
      - Invites
  /v1/api/invites/{code}/complete:
    post:
      description: "In the case of a user invite, this resource will assign the new\
        \ service to the existing user and disables the invite. <br>In the case of\
        \ a service invite, this resource will create a new service, assign gateway\
        \ account ids (if provided) and also creates a new user and assign to the\
        \ service<br>The response contains the user and the service id's affected\
        \ as part of the invite completion in addition to the invite"
      operationId: completeInvite
      parameters:
      - example: d02jddeib0lqpsir28fbskg9v0rv
        in: path
        name: code
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CompleteInviteRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteInviteResponse'
          description: OK
        "404":
          description: Not found
      summary: Completes the invite by creating user/service and invalidating the
        invite code
      tags:
      - Invites
  /v1/api/invites/{code}/reprovision-otp:
    post:
      operationId: reprovisionOtp
      parameters:
      - example: d02jddeib0lqpsir28fbskg9v0rv
        in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
          description: OK
        "404":
          description: Not found
      summary: Re-provision otp secret key for the invite
      tags:
      - Invites
  /v1/api/invites/{code}/send-otp:
    post:
      operationId: sendOtp
      parameters:
      - example: d02jddeib0lqpsir28fbskg9v0rv
        in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No content
        "400":
          description: Invalid payload
        "404":
          description: Not found
        "412":
          description: Precondition failed
      summary: Sends otp verification code to the phone number registered in the invite
      tags:
      - Invites
...
  /v2/api/invites/otp/validate:
    post:
      operationId: validateOtpKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteValidateOtpRequest'
      responses:
        "204":
          description: No content
        "404":
          description: Invite not found
        "422":
          description: Missing required fields or invalid values
      summary: Validates OTP for the invite
      tags:
      - Invites
components:
  schemas:
    CompleteInviteRequest:
      type: object
      properties:
        second_factor:
          type: string
          enum:
          - SMS
          - APP
    CompleteInviteResponse:
      type: object
      properties:
        invite:
          $ref: '#/components/schemas/Invite'
        service_external_id:
          type: string
          example: 89wi6il2364328
        user_external_id:
          type: string
          example: 287cg75v3737
...
    Invite:
      type: object
      properties:
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        attempt_counter:
          type: integer
          format: int32
          example: 0
        disabled:
          type: boolean
          example: false
        email:
          type: string
          example: example@example.gov.uk
        expired:
          type: boolean
          example: false
        inviteLink:
          type: string
          writeOnly: true
        otp_key:
          type: string
          example: ABC123
        password_set:
          type: boolean
          example: false
        role:
          type: string
          example: view-only
        telephone_number:
          type: string
          example: +440787654534
        type:
          type: string
          example: service
        user_exist:
          type: boolean
    InviteServiceRequest:
      type: object
      properties:
        email:
          type: string
          example: example@example.gov.uk
        password:
          type: string
          example: a-password
        telephone_number:
          type: string
          example: +440787654534
      required:
      - email
    InviteUserRequest:
      type: object
      properties:
        email:
          type: string
          example: example@example.gov.uk
        role_name:
          type: string
          example: view-only
        sender:
          type: string
          description: User external ID
          example: d0wksn12nklsdf1nd02nd9n2ndk
        service_external_id:
          type: string
          example: dj2jkejke32jfhh3
      required:
      - email
      - role_name
      - sender
      - service_external_id
    InviteValidateOtpRequest:
      type: object
      properties:
        code:
          type: string
          example: d02jddeib0lqpsir28fbskg9v0rv
          maxLength: 255
        otp:
          type: integer
          format: int32
          example: 123456
      required:
      - code
      - otp
...
    Link:
      type: object
      properties:
        href:
          type: string
          example: https://an.example.link
        method:
          type: string
          example: GET
        rel:
          type: string
          example: self
...